+++ p2panda-rs/src/document/mod.rs
@@ -144,7 +144,7 @@ pub mod error;
 pub mod materialization;
 pub mod traits;
 
-pub use document::{Document, DocumentBuilder, IsDeleted, IsEdited};
+pub use document::{Document, DocumentBuilder};
 pub use document_id::DocumentId;
 pub use document_view::DocumentView;
 pub use document_view_fields::{DocumentViewFields, DocumentViewValue};